Try to split models.py file in several files
to have an more readable and maintanable app
issue #388
