+++ src/member/models/Address.py
@@ -1,0 +1,68 @@
+from django.db import models
+from django.utils.translation import ugettext_lazy as _
+
+class Address(models.Model):
+
+    class Meta:
+        verbose_name_plural = _('addresses')
+
+    # Member address information
+    number = models.PositiveIntegerField(
+        verbose_name=_('street number'),
+        blank=True,
+        null=True,
+    )
+
+    street = models.CharField(
+        max_length=100,
+        verbose_name=_('street')
+    )
+
+    # Can look like B02 so can't be an IntegerField
+    apartment = models.CharField(
+        max_length=10,
+        verbose_name=_('apartment'),
+        blank=True,
+        null=True,
+    )
+
+    floor = models.IntegerField(
+        verbose_name=_('floor'),
+        blank=True,
+        null=True,
+    )
+
+    city = models.CharField(
+        max_length=50,
+        verbose_name=_('city')
+    )
+
+    # Montreal postal code look like H3E1C2
+    postal_code = models.CharField(
+        max_length=6,
+        verbose_name=_('postal code')
+    )
+
+    longitude = models.DecimalField(
+        max_digits=9,
+        decimal_places=6,
+        blank=True,
+        null=True,
+    )
+
+    latitude = models.DecimalField(
+        max_digits=9,
+        decimal_places=6,
+        blank=True,
+        null=True,
+    )
+
+    distance = models.DecimalField(
+        max_digits=9,
+        decimal_places=6,
+        blank=True,
+        null=True,
+    )
+
+    def __str__(self):
+        return self.street
