+++ src/member/models/Referencing.py
@@ -1,0 +1,38 @@
+import datetime
+
+from django.db import models
+from django.utils.translation import ugettext_lazy as _
+
+from member.models import Member, Client
+
+class Referencing (models.Model):
+
+    class Meta:
+        verbose_name_plural = _('referents')
+
+    referent = models.ForeignKey('member.Member',
+                                 verbose_name=_('referent'))
+
+    client = models.ForeignKey('member.Client',
+                               verbose_name=_('client'),
+                               related_name='client_referent')
+
+    referral_reason = models.TextField(
+        verbose_name=_("Referral reason")
+    )
+
+    work_information = models.TextField(
+        verbose_name=_('Work information'),
+        blank=True,
+        null=True,
+    )
+
+    date = models.DateField(verbose_name=_("Referral date"),
+                            auto_now=False, auto_now_add=False,
+                            default=datetime.date.today)
+
+    def __str__(self):
+        return "{} {} referred {} {} on {}".format(
+            self.referent.firstname, self.referent.lastname,
+            self.client.member.firstname, self.client.member.lastname,
+            str(self.date))
