+++ src/member/models/__init__.py
@@ -1,0 +1,12 @@
+from member.models import Address
+from member.models import Client_avoid_ingredient
+from member.models import Client_avoid_component
+from member.models import Client_option
+from member.models import Client
+from member.models import ClientScheduledStatus
+from member.models import Contact
+from member.models import Member
+from member.models import Option
+from member.models import Referencing
+from member.models import Restriction
+from member.models import Route
